Kernel: Make power state switch code mostly infallible
This is done by making the power state switch code to rely on a process
that is always alive, initially started during boot.
Other functions are infallible as well - such as the procedure to kill
all user processes, so now they're marked as such.
