+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -808,7 +808,7 @@ executable("Kernel_bin") {
     "Tasks/FinalizerTask.cpp",
     "Tasks/FutexQueue.cpp",
     "Tasks/PerformanceEventBuffer.cpp",
-    "Tasks/PowerStateSwitchTask.cpp",
+    "Tasks/PowerManagementTask.cpp",
     "Tasks/Process.cpp",
     "Tasks/ProcessGroup.cpp",
     "Tasks/ProcessList.cpp",
